+++ src/labbench/util.py
@@ -121,7 +121,6 @@ def show_messages(
 
     logger._screen_handler = logging.StreamHandler()
     logger._screen_handler.setLevel(level)
-    # - %(pathname)s:%(lineno)d'
 
     if colors:
         log_fmt = '\x1b[1;30m{levelname:^7s}\x1b[0m \x1b[32m{asctime}.{msecs:03.0f}\x1b[0m • \x1b[34m{label}:\x1b[0m {message}'